+++ src/content-config.ts
@@ -8,6 +8,12 @@ export interface KeyboardKeys {
   // The key(s) to cycle through the available dictionaries.
   nextDictionary: string[];
 
+  // The key(s) to move the popup up.
+  movePopupUp: string[];
+
+  // The key(s) to move the popup down.
+  movePopupDown: string[];
+
   // The key(s) to entry copy mode.
   startCopy: string[];
 }